# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API документашка'
  version: 1.0.0
paths:
...
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: 'LOGIN - Аутентификация пользователя и выдача пары токенов'
      operationId: 8dcb70df1020986038d098cc08d05dae
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                  example: a@a.a
                password:
                  type: string
                  example: '123'
              type: object
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  user_id:
                    type: integer
                    example: '1'
                  access:
                    type: string
                    example: abcdefg....
                  access_exp:
                    type: integer
                    example: '1712756994'
                type: object
        401:
          description: 'Неверный логин или пароль.'
        422:
          description: 'Валидация не пройдена'
          content:
            application/json:
              schema:
                properties:
                  errors:
                    properties:
                      email:
                        type: arr
                        example: "['error1', 'error2', ...]"
                      password:
                        type: arr
                        example: "['error1', 'error2', ...]"
                    type: object
                type: object
  /api/auth/checkacc:
    get:
      tags:
        - Auth
      summary: 'checkAccess - Проверка access токена'
      operationId: 611e6f116a521cc357b7ddf1a90e0ce2
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  decoded:
                    properties:
                      iss:
                        type: string
                        example: 'http://example.com'
                      sub:
                        type: integer
                        example: '1'
                      per:
                        type: arr
                        example: '[1, 2, 3]'
                      exp:
                        type: integer
                        example: '1712757994'
                      typ:
                        type: string
                        example: AT
                    type: object
                type: object
        401:
          description: 'Валидация провалена'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Токен не найден.'
                type: object
        403:
          description: 'Токен просрочен'
      security:
        -
          jwt: []
  /api/auth/checkref:
    get:
      tags:
        - Auth
      summary: 'checkRefresh - Проверка refresh токена'
      description: 'Refresh токен хранится в куках'
      operationId: 6d7aa40f9eada0b265dc9bf579c0d19d
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  decoded:
                    properties:
                      iss:
                        type: string
                        example: 'http://example.com'
                      sub:
                        type: integer
                        example: '1'
                      per:
                        type: arr
                        example: '[1, 2, 3]'
                      exp:
                        type: integer
                        example: '1715349934'
                      typ:
                        type: string
                        example: RT
                    type: object
                type: object
        401:
          description: 'Валидация провалена'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Токен не найден.'
                type: object
        403:
          description: 'Токен просрочен, либо недействителен'
      security:
        -
          jwt: []
  /api/auth/refresh:
    get:
      tags:
        - Auth
      summary: 'REFRESH - Обновление пары токенов'
      operationId: b655e0f8fa74e3a128cf4548d7fe3a55
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  user_id:
                    type: integer
                    example: '1'
                  access:
                    type: string
                    example: abcdefg....
                  access_exp:
                    type: integer
                    example: '1712756994'
                type: object
        401:
          description: 'Валидация провалена'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Токен не найден.'
                type: object
        403:
          description: 'Refresh токен просрочен, либо недействителен'
      security:
        -
          jwt: []
  /api/auth/logout:
    get:
      tags:
        - Auth
      summary: 'LOGOUT - Удаление refresh токена из куков и белого списка в БД'
      operationId: a8dd056be3e98a38c657f73ed6cdef2d
      responses:
        200:
          description: 'Токен удалён из белого списка БД. Куки очищены.'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Токен успешно удалён из базы и куков.'
                type: object
        401:
          description: 'Валидация провалена. Куки очищены.'
        403:
          description: 'Refresh токен просрочен, либо недействителен. Куки очищены.'
      security:
        -
          jwt: []
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer